fix task_attempt cleanup, refactor async
